Add center and size properties to controllers
